+++ app/src/main/java/pro/inmost/android/visario/domain/usecases/contacts/DeleteContactUseCase.kt
@@ -1,5 +1,5 @@
 package pro.inmost.android.visario.domain.usecases.contacts
 
 interface DeleteContactUseCase {
-    suspend fun delete(id: Int): Result<Unit>
+    suspend fun delete(id: Long): Result<Unit>
 }
